refactor(contract): split portable backend protocol module
